requirements.txt
Django==3.2.5
psycopg2
pandas
openpyxl
gunicorn==21.2.0
djangorestframework==3.14.0
cryptography~=41.0.2